+++ src/bin/bin.ts
@@ -1,12 +1,14 @@
 #!/usr/bin/env node
 
+import type Program from '../lib/program.js'
+
 import path from 'path'
 import fs from 'fs'
 import url from 'url'
 import c from 'chalk'
 import { program } from 'commander'
 import Terminal from '../lib/terminal/terminal.js'
-import Program from '../lib/program.js'
+import * as constants from '../constants.js'
 
 const __filename = url.fileURLToPath(import.meta.url)
 const __dirname = url.fileURLToPath(new URL('.', import.meta.url))
@@ -18,9 +20,6 @@ interface Package {
     description: string
 }
 
-const DEFAULT_CONFIG_FILE = 'soybean.config.js'
-const CONFIG_FILE_NAMES = [ DEFAULT_CONFIG_FILE ]
-
 // ==================================================================
 
 function getConfigLocation(givenLocation: string = ''): [boolean, string, string] {
@@ -28,7 +27,7 @@ function getConfigLocation(givenLocation: string = ''): [boolean, string, string
 
     // Set givenLocation to the default config file name if it is 
     // an optional parameter, such as --experimental-fetch.
-    if (/--.+/.test(givenLocation)) givenLocation = CONFIG_FILE_NAMES[0]
+    if (/--.+/.test(givenLocation)) givenLocation = constants.CONFIGURATION_FILENAMES[0]
 
     // Check whether a file exists
     const exists = (file: string) => {
@@ -43,8 +42,8 @@ function getConfigLocation(givenLocation: string = ''): [boolean, string, string
     if (!isAbs) givenLocation = path.join(process.cwd(), givenLocation) 
 
     if (!path.extname(bName)) {
-        for (let i = 0; i < CONFIG_FILE_NAMES.length; i++) {
-            const filename = path.join(givenLocation, CONFIG_FILE_NAMES[i])
+        for (let i = 0; i < constants.CONFIGURATION_FILENAMES.length; i++) {
+            const filename = path.join(givenLocation, constants.CONFIGURATION_FILENAMES[i])
             if (exists(filename)) return [true, filename, path.relative(process.cwd(), filename)]
         }
     }
@@ -77,9 +76,12 @@ program
         const program: Program = module.exports || module.default || module
         const cwd = path.dirname(relativeLocation)
 
-        if (program instanceof Program === false) Terminal.EXIT_HARD(`Used configuration file must export a program instance created with the "Soybean" function.`)
-
-        program.start(cwd)
+        if (program.$$SOYBEAN_RELEASE && Array.isArray(program.$$SOYBEAN_RELEASE)) {
+            if (constants.RELEASE_VERSION[0] === program.$$SOYBEAN_RELEASE[0]) return program.start(cwd)
+            Terminal.EXIT_HARD(`Conflicting Soybean instance versions. Global: v${constants.RELEASE_VERSION_STRING}, local: ${program.$$SOYBEAN_RELEASE.join('.')}`)
+        }
+        
+        Terminal.EXIT_HARD(`Used configuration file must export a program instance created with the "Soybean" function.`)
 
     })
 
@@ -93,7 +95,7 @@ program
         if (found && !flags.force) return Terminal.EXIT_HARD(`A configuration file already exists in this location. Use -f to force the creation of a new one.`)
 
         const template = path.join(__dirname, '../../templates/soybean.config.js')
-        const destination = found ? actualLocation : path.join(actualLocation, DEFAULT_CONFIG_FILE)
+        const destination = found ? actualLocation : path.join(actualLocation, constants.DEFAULT_CONFIGURATION_FILENAME)
         await fs.promises.copyFile(template, destination)
         
         Terminal.INFO(`Created a file in "${destination}"`)