+++ src/constants.ts
@@ -1,0 +1,14 @@
+
+import fs from 'fs'
+import path from 'path'
+import * as url from 'url'
+const __filename = url.fileURLToPath(import.meta.url)
+const __dirname = url.fileURLToPath(new URL('.', import.meta.url))
+
+// CONFIGURATION
+export const DEFAULT_CONFIGURATION_FILENAME = 'soybean.config.js'
+export const CONFIGURATION_FILENAMES = [ DEFAULT_CONFIGURATION_FILENAME ]
+
+// VERSIONS
+export const RELEASE_VERSION_STRING = (JSON.parse(fs.readFileSync(path.join(__dirname, '../package.json'), 'utf-8')) as Record<string, string>).version
+export const RELEASE_VERSION = RELEASE_VERSION_STRING.split('.').map(x => Number(x)) as [number, number, number]