+++ CHANGELOG.md
@@ -8,7 +8,8 @@ All notable changes to this project will be documented in this file.
 - Added `event.stopPropagation` function allowing to half the event inside a group handler.
 - Added `json.parse` handler.
 - Added `json.stringify` handler.
-- Added `net.fetch` handler.s
+- Added `net.fetch` handler.
+- Added `os.platform` handler.
 
 ### Fixed
 - Fixed a typo in the error message shown from the help command when the user tries to show details about a non-existent command.