+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Added `forIn` handler.
 - Added `forOf` handler.
 - Added `e.del()` method for deleting information stored on the event object.
+- Added `soybean init [file]` command that lets the user create a boilerplate configuration file.
 
 ### Fixed
 - Fixed a typo in the error message shown from the help command when the user tries to show details about a non-existent command.